Move types file into lib dir
