+++ components/templates/DefaultTemplate/DefaultTemplate.tsx
@@ -3,10 +3,10 @@ import Head from 'next/head';
 import Header from '@/components/UI/molecules/Header';
 import SearchBar from '@/components/UI/molecules/SearchBar';
 import useDebounce from '@/pages/hooks/useDebounce';
-import { useFindPokemonSuggestions } from '@/pages/hooks/usePokemon';
 import Container from '@/components/UI/atoms/Container';
 import { styled } from '@/stitches.config';
 import Title from '@/components/UI/atoms/Title';
+import useFindPokemonSuggestions from '@/pages/hooks/usePokemon';
 
 interface Props {
   children?: React.ReactNode;
